scripts/new_submissions/bot-requirements.txt
python-telegram-bot>=13.0,<14.0
pyyaml
